Move fetchTaxonAndSave to own file
